+++ examples/fusion_extraction.py
@@ -24,7 +24,7 @@ from graphix.extraction import get_fusion_network_from_graph
 
 # %%
 # Here we say we want a graph state with 9 nodes and 12 edges.
-# We can obtain resource graph for a measurement pattern by using :code:`nodes, edges = pattern.get_graph()`.
+# We can obtain resource graph for a measurement pattern by using :code:`pattern.extract_graph()`.
 gs = graphix.GraphState()
 nodes = [0, 1, 2, 3, 4, 5, 6, 7, 8]
 edges = [(0, 1), (1, 2), (2, 3), (3, 0), (3, 4), (0, 5), (4, 5), (5, 6), (6, 7), (7, 0), (7, 8), (8, 1)]